Query parser refactoring: delete Parser T-Z
