+++ src/main/java/org/elasticsearch/index/query/WildcardQueryBuilder.java
@@ -19,7 +19,15 @@
 
 package org.elasticsearch.index.query;
 
+import org.apache.lucene.index.Term;
+import org.apache.lucene.search.Query;
+import org.apache.lucene.search.WildcardQuery;
+import org.apache.lucene.util.BytesRef;
+import org.elasticsearch.common.inject.Inject;
 import org.elasticsearch.common.xcontent.XContentBuilder;
+import org.elasticsearch.common.xcontent.XContentParser;
+import org.elasticsearch.index.mapper.MapperService;
+import org.elasticsearch.index.query.support.QueryParsers;
 
 import java.io.IOException;
 
@@ -33,7 +41,7 @@ import java.io.IOException;
  *
  *
  */
-public class WildcardQueryBuilder extends BaseQueryBuilder implements MultiTermQueryBuilder, BoostableQueryBuilder<WildcardQueryBuilder> {
+public class WildcardQueryBuilder extends BaseQueryBuilder implements QueryParser, MultiTermQueryBuilder, BoostableQueryBuilder<WildcardQueryBuilder> {
 
     private final String name;
 
@@ -44,6 +52,14 @@ public class WildcardQueryBuilder extends BaseQueryBuilder implements MultiTermQ
     private String rewrite;
 
     private String queryName;
+    
+    
+    public static final String NAME = "wildcard";
+
+    @Inject
+    public WildcardQueryBuilder() {
+        this(null, null);
+    }
 
     /**
      * Implements the wildcard search query. Supported wildcards are <tt>*</tt>, which
@@ -86,7 +102,7 @@ public class WildcardQueryBuilder extends BaseQueryBuilder implements MultiTermQ
 
     @Override
     public void doXContent(XContentBuilder builder, Params params) throws IOException {
-        builder.startObject(WildcardQueryParser.NAME);
+        builder.startObject(WildcardQueryBuilder.NAME);
         if (boost == -1 && rewrite == null && queryName != null) {
             builder.field(name, wildcard);
         } else {
@@ -104,5 +120,75 @@ public class WildcardQueryBuilder extends BaseQueryBuilder implements MultiTermQ
             builder.endObject();
         }
         builder.endObject();
+    }
+
+    @Override
+    public String[] names() {
+        return new String[]{NAME};
+    }
+
+    @Override
+    public Query parse(QueryParseContext parseContext) throws IOException, QueryParsingException {
+        XContentParser parser = parseContext.parser();
+
+        XContentParser.Token token = parser.nextToken();
+        if (token != XContentParser.Token.FIELD_NAME) {
+            throw new QueryParsingException(parseContext.index(), "[wildcard] query malformed, no field");
+        }
+        String fieldName = parser.currentName();
+        String rewriteMethod = null;
+
+        String value = null;
+        float boost = 1.0f;
+        String queryName = null;
+        token = parser.nextToken();
+        if (token == XContentParser.Token.START_OBJECT) {
+            String currentFieldName = null;
+            while ((token = parser.nextToken()) != XContentParser.Token.END_OBJECT) {
+                if (token == XContentParser.Token.FIELD_NAME) {
+                    currentFieldName = parser.currentName();
+                } else {
+                    if ("wildcard".equals(currentFieldName)) {
+                        value = parser.text();
+                    } else if ("value".equals(currentFieldName)) {
+                        value = parser.text();
+                    } else if ("boost".equals(currentFieldName)) {
+                        boost = parser.floatValue();
+                    } else if ("rewrite".equals(currentFieldName)) {
+                        rewriteMethod = parser.textOrNull();
+                    } else if ("_name".equals(currentFieldName)) {
+                        queryName = parser.text();
+                    } else {
+                        throw new QueryParsingException(parseContext.index(), "[wildcard] query does not support [" + currentFieldName + "]");
+                    }
+                }
+            }
+            parser.nextToken();
+        } else {
+            value = parser.text();
+            parser.nextToken();
+        }
+
+        if (value == null) {
+            throw new QueryParsingException(parseContext.index(), "No value specified for prefix query");
+        }
+
+        BytesRef valueBytes;
+        MapperService.SmartNameFieldMappers smartNameFieldMappers = parseContext.smartFieldMappers(fieldName);
+        if (smartNameFieldMappers != null && smartNameFieldMappers.hasMapper()) {
+            fieldName = smartNameFieldMappers.mapper().names().indexName();
+            valueBytes = smartNameFieldMappers.mapper().indexedValueForSearch(value);
+        } else {
+            valueBytes = new BytesRef(value);
+        }
+
+        WildcardQuery wildcardQuery = new WildcardQuery(new Term(fieldName, valueBytes));
+        QueryParsers.setRewriteMethod(wildcardQuery, rewriteMethod);
+        wildcardQuery.setRewriteMethod(QueryParsers.parseRewriteMethod(rewriteMethod));
+        wildcardQuery.setBoost(boost);
+        if (queryName != null) {
+            parseContext.addNamedQuery(queryName, wildcardQuery);
+        }
+        return wildcardQuery;
     }
 }
