+++ src/main/java/org/elasticsearch/index/query/functionscore/DecayFunctionParser.java
@@ -423,16 +423,16 @@ public abstract class DecayFunctionParser implements ScoreFunctionParser {
             super(CombineFunction.MULT);
             this.mode = mode;
             if (userSuppiedScale <= 0.0) {
-                throw new ElasticsearchIllegalArgumentException(FunctionScoreQueryParser.NAME + " : scale must be > 0.0.");
+                throw new ElasticsearchIllegalArgumentException(FunctionScoreQueryBuilder.NAME + " : scale must be > 0.0.");
             }
             if (decay <= 0.0 || decay >= 1.0) {
-                throw new ElasticsearchIllegalArgumentException(FunctionScoreQueryParser.NAME
+                throw new ElasticsearchIllegalArgumentException(FunctionScoreQueryBuilder.NAME
                         + " : decay must be in the range [0..1].");
             }
             this.scale = func.processScale(userSuppiedScale, decay);
             this.func = func;
             if (offset < 0.0d) {
-                throw new ElasticsearchIllegalArgumentException(FunctionScoreQueryParser.NAME + " : offset must be > 0.0");
+                throw new ElasticsearchIllegalArgumentException(FunctionScoreQueryBuilder.NAME + " : offset must be > 0.0");
             }
             this.offset = offset;
         }
